Adds features to help in content selection.
First, a new table is added to the database, providing the means to
specify what section a specific question came from.

Second, the session description transmitted when creating a session
has been modified to be able to include the required data.

Last, the web site has been updated to include the required forms.